refactor SwapExactAmountIn
